requirements.txt
kopf==1.37.0
kubernetes==28.1.0
prometheus-client==0.19.0
redis==5.0.1
google-cloud-pubsub==2.18.4
pytest==7.4.3
requests==2.31.0
